backend_api/pyproject.toml
[project]
name = "backend-api"
version = "0.1.0"
description = "지지저항Lab MVP backend scaffold"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "alembic>=1.16.0,<2.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
test = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.0.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
